+++ lib/saxophone/sax_config.rb
@@ -1,10 +1,10 @@
-require "sax-machine/config/sax_element_value"
-require "sax-machine/config/sax_attribute"
-require "sax-machine/config/sax_element"
-require "sax-machine/config/sax_collection"
-require "sax-machine/config/sax_ancestor"
+require "saxophone/config/sax_element_value"
+require "saxophone/config/sax_attribute"
+require "saxophone/config/sax_element"
+require "saxophone/config/sax_collection"
+require "saxophone/config/sax_ancestor"
 
-module SAXMachine
+module Saxophone
   class SAXConfig
     attr_accessor :top_level_elements, :top_level_attributes, :top_level_element_value, :collection_elements, :ancestors