+++ spec/spec_helper.rb
@@ -13,8 +13,8 @@ begin
 rescue LoadError
 end
 
-require File.expand_path(File.dirname(__FILE__) + '/../lib/sax-machine')
-SAXMachine.handler = ENV['HANDLER'].to_sym if ENV['HANDLER']
+require File.expand_path(File.dirname(__FILE__) + '/../lib/saxophone')
+Saxophone.handler = ENV['HANDLER'].to_sym if ENV['HANDLER']
 
 RSpec.configure do |config|
   config.run_all_when_everything_filtered = true